// src/types/userTypes.ts
export type UserDataProp = {
  address: string;
  _id: string;
  email: string;
  phone: string[];
  isAdmin: boolean;
  profileCreated: boolean;
  createdAt: string;
  updatedAt: string;
  __v: number;
  additionalInformation: string;
  name: string;
  location:string;
  firstname: string;
  lastname: string;
  profilePhoto: string;
  postal_code: string;
};